Start work (not finish) on precise stats. Are they even needed?

+++ OnTheFlyStats.Test/ExtensionsTest.cs
@@ -10,8 +10,8 @@ namespace OnTheFlyStats.Test
     public class ExtensionsTest
     {
         [Theory]
-        [InlineData(new double[] { 1.0, 2, 3, 3.14, 4, 1, -1, 7, -11, 15 }, -11, 15, 42.07)]
-        public void StatsExtensionCalculatesProperly(double[] input, double min, double max, double variance)
+        [InlineData(new decimal[] { 1.0m, 2, 3, 3.14, 4, 1, -1, 7, -11m, 15m }, -11m, 15m, 42.07m)]
+        public void StatsExtensionCalculatesProperly(decimal[] input, decimal min, decimal max, decimal variance)
         {
             var result = input.AsEnumerable().Stats();
             Assert.Equal(input.Length, result.N);
